Zarr sharding codec

The Zarr sharding codec is not a true codec in the sense that it does not encode or decode - that is left up to regular codec defined inside this "codec" configuration. This implementation can read from a store using sharding, writing is not supported.

Super classes

zarr::zarr_extension -> zarr::zarr_codec -> zarr_codec_sharding

Methods

Public methods

zarr_codec_sharding$new()

zarr_codec_sharding$copy()

Inherited methods

zarr::zarr_codec$decode()

zarr::zarr_codec$encode()

zarr::zarr_codec$metadata_fragment()

zarr::zarr_codec$print()

Method new()

Create a new "crc32c" codec object.

Usage

zarr_codec_sharding$new(configuration)

Arguments

configuration

Optional. A list with the configuration parameters for this codec but since this codec doesn't have any the argument is always ignored.

Returns

An instance of this class.

Method copy()

Create a new, independent copy of this codec.

Usage

zarr_codec_sharding$copy()

Returns

An instance of zarr_codec_sharding.