zarr: Native and Extensible R Driver for 'Zarr'

The 'Zarr' specification is widely used to build libraries for the storage and retrieval of n-dimensional array data from data stores ranging from local file systems to the cloud. This package is a native 'Zarr' implementation in R with support for all required features of 'Zarr' version 3. It is designed to be extensible such that new stores, codecs and extensions can be added easily.

See also

Author

Maintainer: Patrick Van Laake patrick@vanlaake.net [copyright holder]